Locally deleted GIT_DIR and GIT_WORK_TREE.
Git::Repository uses GIT_DIR and GIT_WORK_TREE to determine the path to the git repository when those environment variables are present. This however poses problems here, when those variables point to a different repository then the one the file to analyze belongs to, or when they use relative paths. To force Git::Repository to derive the git repository's path from the file path, we thus locally delete GIT_DIR and GIT_WORK_TREE.